Switched from double to single floating point precision

+++ src/IOParser.cpp
@@ -38,21 +38,21 @@ void::IOParser::OpenOutFile() {
 void::IOParser::WriteStateToOutFile(Particle* state, int n) {
 
     // create temporary arrays for writing
-    double* arr_x = new double[n];
-    double* arr_y = new double[n];
-    double* arr_colour = new double[n];
+    float* arr_x = new float[n];
+    float* arr_y = new float[n];
+    float* arr_colour = new float[n];
 
     // fill temp arrays from particle structs in state
     for (int i = 0; i < n; i++) {
         arr_x[i] = state[i].x;
         arr_y[i] = state[i].y;
-        arr_colour[i] = (double)state[i].colour;
+        arr_colour[i] = (float)state[i].colour;
     }
 
     // write coords, and colours
-    io_out_disp_file.write((char *)arr_x, sizeof(double) * n);
-    io_out_disp_file.write((char *)arr_y, sizeof(double) * n);
-    io_out_disp_file.write((char *)arr_colour, sizeof(double) * n);
+    io_out_disp_file.write((char *)arr_x, sizeof(float) * n);
+    io_out_disp_file.write((char *)arr_y, sizeof(float) * n);
+    io_out_disp_file.write((char *)arr_colour, sizeof(float) * n);
 
     // flush output stream, but keep file open
     io_out_disp_file.flush();
